+++ src/latlng.h
@@ -16,12 +16,12 @@ class LatLng : public Napi::ObjectWrap<LatLng> {
 
     S2LatLng Get();
 
-
   private:
     Napi::Value ToString(const Napi::CallbackInfo &info);
     Napi::Value Normalized(const Napi::CallbackInfo &info);
     Napi::Value Latitude(const Napi::CallbackInfo &info);
     Napi::Value Longitude(const Napi::CallbackInfo &info);
+    Napi::Value ApproxEquals(const Napi::CallbackInfo &info);
     S2LatLng s2latlng;
 };