+++ test.sh
@@ -5,11 +5,10 @@ set -ex
 git submodule init
 git submodule update --remote
 
-
 # test node 10
 NODE_10_IMAGE=test10
 NODE_10_DOCKERFILE=./docker/Dockerfile.node10.test
-docker build -t $NODE_10_IMAGE -f ./docker/Dockerfile.node10.test . && docker run $NODE_10_IMAGE:latest
+docker build -t $NODE_10_IMAGE -f $NODE_10_DOCKERFILE . && docker run $NODE_10_IMAGE:latest
 
 # test node 12
 NODE_12_IMAGE=test12